Add BottomBanner component to encapsulate logic
