+++ src/web/components/BottomBanner/ReaderRevenueBanner.tsx
